<Update> Last minute fix warnings

+++ FlexEngine/src/FlexEngine/Renderer/scenecamsorter.cpp
@@ -125,8 +125,8 @@ namespace FlexEngine
     void SceneCamSorter::RemoveCameraEntities()
     {
       m_cameraEntities.clear();
-      m_currMainID = -1;
-      m_currEditorID = -1;
+      m_currMainID = static_cast<uint64_t>(-1);
+      m_currEditorID = static_cast<uint64_t>(-1);
     }
 
     /*!***************************************************************************
